More renames and additional READMEs/LICENSEs/etc.
