+++ web_generator/bin/scrape_mdn.dart
@@ -77,7 +77,7 @@ Future<void> main(List<String> args) async {
 
   const encoder = JsonEncoder.withIndent('  ');
 
-  final file = File('third_party/mdn/mdn.json');
+  final file = File('../third_party/mdn/mdn.json');
   final json = {
     '__meta__': {
       'source': '[MDN Web Docs]($mdnUrl)',
