--- web_generator/generator/README.md
@@ -1,19 +1,0 @@
-The bindings generator for `package:web`.
-
-This tool is written in Dart, compiled to JavaScript, and run on Node.
-
-## Generating the bindings
-
-Run:
-
-```
-dart update_bindings.dart
-```
-
-## Update to the latest Web IDL versions and regenerate
-
-Run:
-
-```
-dart update_bindings.dart --update
-```
