+++ web_generator/lib/src/banned_names.dart
