+++ web_generator/lib/src/dart_main.dart
