+++ web_generator/lib/src/generate_bindings.dart
