+++ web_generator/lib/src/package.json
