+++ web_generator/lib/src/singletons.dart
