+++ web_generator/lib/src/translator.dart
