+++ web_generator/mono_pkg.yaml
