+++ web_generator/test/type_union_test.dart
@@ -4,7 +4,7 @@
 
 import 'package:test/test.dart';
 
-import '../generator/type_union.dart';
+import 'package:web_generator/src/type_union.dart';
 
 void main() {
   test('Non-JS types', () {
